+++ geniusyield-server-lib/src/GeniusYield/Server/Dex/PartialOrder.hs
@@ -5,7 +5,6 @@ module GeniusYield.Server.Dex.PartialOrder (
   poiToOrderInfo,
   PodServerException (..),
   PodOrderNotFound (..),
-  ErrDescription (..),
 ) where
 
 import Data.Aeson (ToJSON (..))
@@ -33,7 +32,6 @@ import RIO.Map qualified as Map
 import RIO.NonEmpty qualified as NonEmpty
 import RIO.Text qualified as T
 import Servant
-import Servant.Checked.Exceptions
 
 -- | Number of orders that we at most allow to be filled in a single transaction.
 maxFillOrders ∷ GYNatural
@@ -52,8 +50,10 @@ data PodOrderNotFound = PodOrderNotFound
   deriving (Eq, Show, Generic)
   deriving anyclass (Exception, ToJSON)
 
-instance ErrStatus PodOrderNotFound where
-  toErrStatus _ = status404
+instance Swagger.ToSchema PodOrderNotFound where
+  declareNamedSchema =
+    Swagger.genericDeclareNamedSchema Swagger.defaultSchemaOptions
+      & addSwaggerDescription (toErrDescription PodOrderNotFound)
 
 class ErrDescription e where
   toErrDescription ∷ e → Text
@@ -320,6 +320,23 @@ instance Swagger.ToSchema FillOrderParameters where
       & addSwaggerDescription "Fill order(s) request parameters."
       & addSwaggerExample (toJSON $ FillOrderParameters {fopAddresses = pure "addr_test1qrsuhwqdhz0zjgnf46unas27h93amfghddnff8lpc2n28rgmjv8f77ka0zshfgssqr5cnl64zdnde5f8q2xt923e7ctqu49mg5", fopChangeAddress = Just (ChangeAddress "addr_test1qrsuhwqdhz0zjgnf46unas27h93amfghddnff8lpc2n28rgmjv8f77ka0zshfgssqr5cnl64zdnde5f8q2xt923e7ctqu49mg5"), fopCollateral = Just "4293386fef391299c9886dc0ef3e8676cbdbc2c9f2773507f1f838e00043a189#1", fopOrderReferencesWithAmount = ("0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e1ea747#0", 100) :| [("0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e144444#0", 100)]})
 
+newtype BotFillOrderParameters = BotFillOrderParameters
+  { bfopOrderReferencesWithAmount ∷ NonEmpty (GYTxOutRef, GYNatural)
+  }
+  deriving stock (Show, Generic)
+  deriving
+    (FromJSON, ToJSON)
+    via CustomJSON '[FieldLabelModifier '[StripPrefix BotFillOrderReqPrefix, CamelToSnake]] BotFillOrderParameters
+
+type BotFillOrderReqPrefix ∷ Symbol
+type BotFillOrderReqPrefix = "bfop"
+
+instance Swagger.ToSchema BotFillOrderParameters where
+  declareNamedSchema =
+    Swagger.genericDeclareNamedSchema Swagger.defaultSchemaOptions {Swagger.fieldLabelModifier = dropSymbolAndCamelToSnake @BotFillOrderReqPrefix}
+      & addSwaggerDescription "Fill order(s) request parameters specialized towards configured bot."
+      & addSwaggerExample (toJSON $ BotFillOrderParameters {bfopOrderReferencesWithAmount = ("0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e1ea747#0", 100) :| [("0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e144444#0", 100)]})
+
 type FillOrderResPrefix ∷ Symbol
 type FillOrderResPrefix = "fotd"
 
@@ -344,7 +361,7 @@ type CommonCollateralText ∷ Symbol
 type CommonCollateralText = "Note that if \"collateral\" field is not provided, then framework would try to pick collateral UTxO on it's own and in that case would also be free to spend it (i.e., would be made available to coin balancer)."
 
 type CommonSignText ∷ Symbol
-type CommonSignText = "It uses the signing key from configuration to compute for wallet address. If collateral is specified in the configuration, then it would be used for."
+type CommonSignText = "This endpoint would also sign & submit the built transaction. It uses the signing key from configuration to compute for wallet address. If collateral is specified in the configuration, then it would be used for."
 
 type OrdersAPI =
   Summary "Build transaction to create order"
@@ -354,7 +371,7 @@ type OrdersAPI =
     :> ReqBody '[JSON] PlaceOrderParameters
     :> Post '[JSON] PlaceOrderTransactionDetails
     :<|> Summary "Create an order"
-      :> Description ("Create an order. This endpoint would also sign & submit the built transaction. " `AppendSymbol` CommonSignText `AppendSymbol` " \"stakeAddress\" field from configuration, if provided, is used to place order at a mangled address.")
+      :> Description ("Create an order. " `AppendSymbol` CommonSignText `AppendSymbol` " \"stakeAddress\" field from configuration, if provided, is used to place order at a mangled address.")
       :> ReqBody '[JSON] BotPlaceOrderParameters
       :> Post '[JSON] PlaceOrderTransactionDetails
     :<|> Summary "Build transaction to cancel order(s)"
@@ -364,7 +381,7 @@ type OrdersAPI =
       :> ReqBody '[JSON] CancelOrderParameters
       :> Post '[JSON] CancelOrderTransactionDetails
     :<|> Summary "Cancel order(s)"
-      :> Description ("Cancel order(s). This endpoint would also sign & submit the built transaction. " `AppendSymbol` CommonSignText)
+      :> Description ("Cancel order(s). " `AppendSymbol` CommonSignText)
       :> ReqBody '[JSON] BotCancelOrderParameters
       :> Delete '[JSON] CancelOrderTransactionDetails
     :<|> Summary "Get order(s) details"
@@ -376,13 +393,17 @@ type OrdersAPI =
       :> Description "Get details of an order using it's unique NFT token. Note that each order is identified uniquely by an associated NFT token which can then later be used to retrieve it's details across partial fills."
       :> "details"
       :> Capture "nft-token" GYAssetClass
-      :> Throws PodOrderNotFound
-      :> Get '[JSON] OrderInfoDetailed
+      :> UVerb 'GET '[JSON] '[WithStatus 200 OrderInfoDetailed, WithStatus 404 PodOrderNotFound]
     :<|> Summary "Build transaction to fill order(s)"
       :> Description ("Build a transaction to fill order(s). " `AppendSymbol` CommonCollateralText)
       :> "tx"
       :> "build-fill"
       :> ReqBody '[JSON] FillOrderParameters
+      :> Post '[JSON] FillOrderTransactionDetails
+    :<|> Summary "Build transaction to fill order(s)"
+      :> Description ("Build a transaction to fill order(s). " `AppendSymbol` CommonSignText)
+      :> "fill"
+      :> ReqBody '[JSON] BotFillOrderParameters
       :> Post '[JSON] FillOrderTransactionDetails
 
 handleOrdersApi ∷ Ctx → ServerT OrdersAPI IO
@@ -394,6 +415,7 @@ handleOrdersApi ctx =
     :<|> handleOrdersDetails ctx
     :<|> handleOrderDetails ctx
     :<|> handleFillOrders ctx
+    :<|> handleFillOrdersAndSignSubmit ctx
 
 handlePlaceOrder ∷ Ctx → PlaceOrderParameters → IO PlaceOrderTransactionDetails
 handlePlaceOrder ctx@Ctx {..} pops@PlaceOrderParameters {..} = do
@@ -477,14 +499,14 @@ handleCancelOrdersAndSignSubmit ctx BotCancelOrderParameters {..} = do
   -- Though transaction id would be same, but we are returning it again, just in case...
   pure $ details {cotdTransactionId = txId, cotdTransaction = signedTx}
 
-handleOrderDetails ∷ Ctx → GYAssetClass → IO (Envelope '[PodOrderNotFound] OrderInfoDetailed)
+handleOrderDetails ∷ Ctx → GYAssetClass → IO (Union '[WithStatus 200 OrderInfoDetailed, WithStatus 404 PodOrderNotFound])
 handleOrderDetails ctx@Ctx {..} ac = do
   logInfo ctx $ "Getting order details for NFT token: " +|| ac ||+ ""
   let porefs = dexPORefs ctxDexInfo
   os ← runQuery ctx $ fmap poiToOrderInfoDetailed <$> orderByNft porefs ac
   case os of
-    Nothing → throwIO PodOrderNotFound
-    Just o → pureSuccEnvelope o
+    Nothing → throwIO PodOrderNotFound -- We could use `respond` here as well but then as it would not have @application/json@ header, it would not be caught by our @errorJsonWrapMiddleware@.
+    Just o → respond (WithStatus @200 o)
 
 handleOrdersDetails ∷ Ctx → [GYAssetClass] → IO [OrderInfoDetailed]
 handleOrdersDetails ctx@Ctx {..} acs = do
@@ -539,3 +561,13 @@ handleFillOrders ctx@Ctx {..} fops@FillOrderParameters {..} = do
         takerFee =
           Map.foldlWithKey' (\acc ac amt → acc <> valueSingleton ac (roundFunctionForPOCVersion overallPocVersion $ toRational amt * rationalToGHC takerFeeRatio)) mempty takerACWithAmt
      in takerFee
+
+handleFillOrdersAndSignSubmit ∷ Ctx → BotFillOrderParameters → IO FillOrderTransactionDetails
+handleFillOrdersAndSignSubmit ctx BotFillOrderParameters {..} = do
+  logInfo ctx "Filling order(s) and signing & submitting the transaction."
+  ctxAddr ← addressToBech32 <$> resolveCtxAddr ctx
+  details ← handleFillOrders ctx $ FillOrderParameters {fopAddresses = pure ctxAddr, fopChangeAddress = Just (ChangeAddress ctxAddr), fopCollateral = ctxCollateral ctx, fopOrderReferencesWithAmount = bfopOrderReferencesWithAmount}
+  signedTx ← handleTxSign ctx $ fotdTransaction details
+  txId ← handleTxSubmit ctx signedTx
+  -- Though transaction id would be same, but we are returning it again, just in case...
+  pure $ details {fotdTransactionId = txId, fotdTransaction = signedTx}